pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dg2000-trigger"
version = "0.2.0"
description = "DG2000 波形发生器触发与循环输出控制（PySide6 + PyVISA）"
requires-python = ">=3.9"
dependencies = [
  "pyvisa>=1.14.0",
  "pyvisa-py>=0.7.2",
  "pyusb>=1.3.1",
  "PySide6-Essentials>=6.7.0",
]

[project.optional-dependencies]
build = ["pyinstaller>=6.6"]

[project.scripts]
dg2000-trigger = "dg2000_trigger.app:main"

[tool.setuptools.packages.find]
include = ["dg2000_trigger*"]
